Added BaseSongAdapter.kt
