+++ app/src/main/java/com/plcoding/spotifycloneyt/adapters/BaseSongAdapter.kt
@@ -1,0 +1,54 @@
+package com.plcoding.spotifycloneyt.adapters
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.AsyncListDiffer
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.RecyclerView
+import com.plcoding.spotifycloneyt.R
+import com.plcoding.spotifycloneyt.data.entities.Song
+import kotlinx.android.synthetic.main.list_item.view.*
+
+abstract class BaseSongAdapter(
+    private val layoutId: Int
+) : RecyclerView.Adapter<BaseSongAdapter.SongViewHolder>() {
+
+    class SongViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
+
+    protected val diffCallback = object : DiffUtil.ItemCallback<Song>() {
+        override fun areItemsTheSame(oldItem: Song, newItem: Song): Boolean {
+            return oldItem.mediaId == newItem.mediaId
+        }
+
+        override fun areContentsTheSame(oldItem: Song, newItem: Song): Boolean {
+            return oldItem.hashCode() == newItem.hashCode()
+        }
+    }
+
+    protected abstract val differ: AsyncListDiffer<Song>
+
+    var songs: List<Song>
+        get() = differ.currentList
+        set(value) = differ.submitList(value)
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SongViewHolder {
+        return SongViewHolder(
+            LayoutInflater.from(parent.context).inflate(
+                layoutId,
+                parent,
+                false
+            )
+        )
+    }
+
+    protected var onItemClickListener: ((Song) -> Unit)? = null
+
+    fun setItemClickListener(listener: (Song) -> Unit) {
+        onItemClickListener = listener
+    }
+
+    override fun getItemCount(): Int {
+        return songs.size
+    }
+}
