+++ app/src/main/java/com/plcoding/spotifycloneyt/adapters/SongAdapter.kt
@@ -14,35 +14,9 @@ import javax.inject.Inject
 
 class SongAdapter @Inject constructor(
     private val glide: RequestManager
-) : RecyclerView.Adapter<SongAdapter.SongViewHolder>() {
+) : BaseSongAdapter(R.layout.list_item) {
 
-    class SongViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
-
-    private val diffCallback = object : DiffUtil.ItemCallback<Song>() {
-        override fun areItemsTheSame(oldItem: Song, newItem: Song): Boolean {
-            return oldItem.mediaId == newItem.mediaId
-        }
-
-        override fun areContentsTheSame(oldItem: Song, newItem: Song): Boolean {
-            return oldItem.hashCode() == newItem.hashCode()
-        }
-    }
-
-    private val differ = AsyncListDiffer(this, diffCallback)
-
-    var songs: List<Song>
-        get() = differ.currentList
-        set(value) = differ.submitList(value)
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SongViewHolder {
-        return SongViewHolder(
-            LayoutInflater.from(parent.context).inflate(
-                R.layout.list_item,
-                parent,
-                false
-            )
-        )
-    }
+    override val differ = AsyncListDiffer(this, diffCallback)
 
     override fun onBindViewHolder(holder: SongViewHolder, position: Int) {
         val song = songs[position]
@@ -59,15 +33,6 @@ class SongAdapter @Inject constructor(
         }
     }
 
-    private var onItemClickListener: ((Song) -> Unit)? = null
-
-    fun setOnItemClickListener(listener: (Song) -> Unit) {
-        onItemClickListener = listener
-    }
-
-    override fun getItemCount(): Int {
-        return songs.size
-    }
 }
