+++ app/src/main/java/com/plcoding/spotifycloneyt/adapters/SwipeSongAdapter.kt
@@ -1,0 +1,44 @@
+package com.plcoding.spotifycloneyt.adapters
+
+import androidx.recyclerview.widget.AsyncListDiffer
+import com.plcoding.spotifycloneyt.R
+import kotlinx.android.synthetic.main.swipe_item.view.*
+
+class SwipeSongAdapter : BaseSongAdapter(R.layout.list_item) {
+
+    override val differ = AsyncListDiffer(this, diffCallback)
+
+    override fun onBindViewHolder(holder: SongViewHolder, position: Int) {
+        val song = songs[position]
+        holder.itemView.apply {
+            val text = "${song.title} - ${song.subtitle}"
+            tvPrimary.text = text
+
+            setOnClickListener {
+                onItemClickListener?.let { click ->
+                    click(song)
+                }
+            }
+        }
+    }
+
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
